add authentication

+++ base/views.py
@@ -1,6 +1,10 @@
 from django.shortcuts import render, redirect
-# from django.http import HttpResponse
+from django.contrib import messages
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
 from .models import Room, Topic
 from .forms import RoomForm
 
@@ -9,6 +13,33 @@ from .forms import RoomForm
 #     {'id':2, 'name' : 'Profile settings '},
 #     {'id':3, 'name' : 'Feature 1'},
 # ]
+
+
+def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = user.objects.get(username=username)
+        except:
+            messages.error(request, 'user does not exist!')
+        
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username OR Password Does not exist')
+    context = {}
+    return render(request, 'base/login_registration.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
 
 def home(request):
 
@@ -32,6 +63,8 @@ def room(request, pk):
     context = {'room' : room}
     return render(request, 'base/room.html', context)
 
+
+@login_required(login_url='/login') #can create a job posting only if they are users
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -42,9 +75,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('<h3>You do not have the access to update this job details!<h3>')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -55,8 +93,13 @@ def updateRoom(request, pk):
     context = {'form' : form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('<h3>You do not have the access to delete this job details!<h3>')
+
     if request.method == 'POST' :
         room.delete()
         return redirect('home')
